refactor: rename file
Module related to wsl was mistakenly called "wls". Fixed name.
